feat!: use regular custom element i/o built-in https://github.com/WebKit/standards-positions/issues/97
